Better dbt seeding in tests (#213)

+++ integration_tests/features/environment.py
@@ -2,8 +2,9 @@ import os
 
 
 def before_all(context):
-    os.system("mkdir mock/temp")
-    os.system("dbt seed --profiles-dir mock --project-dir mock")
+    if not hasattr(context, "temp_dir"):
+        os.system("mkdir mock/temp")
+        os.system("dbt seed --profiles-dir mock --project-dir mock")
 
 
 def after_all(context):
@@ -13,4 +14,5 @@ def after_all(context):
 
 
 def after_scenario(context, scenario):
-    os.system("rm -rf mock/temp/*")
+    if not hasattr(context, "temp_dir"):
+        os.system("rm -rf mock/temp/*")
